HS-446: Move traps feature to it's own module
