+++ minion/pom.xml
@@ -25,7 +25,7 @@
         <module>minion-observability</module>
         <module>minion-taskset</module>
         <module>plugins</module>
-        <module>traps-listener</module>
+        <module>traps</module>
         <module>minion-features</module>
         <module>docker-assembly</module>
         <module>assembly</module>
