--- minion/traps-listener/src/main/java/org/opennms/horizon/minion/traps/TrapInformationWrapper.java
@@ -1,76 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2022 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2022 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.horizon.minion.traps;
-
-import org.opennms.horizon.grpc.traps.contract.TrapDTO;
-import org.opennms.horizon.shared.snmp.SnmpException;
-import org.opennms.horizon.shared.snmp.traps.TrapInformation;
-
-import java.net.InetAddress;
-import java.util.Objects;
-
-public class TrapInformationWrapper {
-
-    private final TrapInformation trapInformation;
-
-    private final TrapDTO trapDTO;
-
-    private InetAddress trapAddress;
-
-    public TrapInformationWrapper(TrapInformation trapInformation) throws SnmpException {
-        this.trapInformation = Objects.requireNonNull(trapInformation);
-        this.trapDTO = null;
-        trapInformation.validate(); // Before this was at ProcessQueueProcessor which does not exist anymore
-    }
-
-    public TrapInformationWrapper(TrapDTO trapDTO) {
-        this.trapDTO = trapDTO;
-        this.trapInformation = null;
-    }
-
-    public TrapInformation getTrapInformation() {
-        return trapInformation;
-    }
-
-    public TrapDTO getTrapDTO() {
-        return trapDTO;
-    }
-
-
-    public InetAddress getTrapAddress() {
-        if (trapInformation != null) {
-            return getTrapInformation().getTrapAddress();
-        }
-        return trapAddress;
-    }
-
-    public void setTrapAddress(InetAddress trapAddress) {
-        this.trapAddress = trapAddress;
-    }
-}
