+++ minion/traps/features/pom.xml
@@ -1,27 +1,24 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>minion</artifactId>
+        <artifactId>traps</artifactId>
         <groupId>org.opennms.horizon.minion</groupId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>minion-features</artifactId>
+    <artifactId>traps-features</artifactId>
     <packaging>feature</packaging>
 
-    <name>OpenNms Horizon :: Minion :: Features</name>
+    <name>OpenNms Horizon :: Minion :: Traps :: Features</name>
 
     <dependencies>
-
         <dependency>
             <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>grpc-features</artifactId>
+            <artifactId>traps-listener</artifactId>
             <version>${project.version}</version>
-            <classifier>features</classifier>
-            <type>xml</type>
         </dependency>
         <dependency>
             <groupId>org.opennms.horizon.minion</groupId>
@@ -32,14 +29,14 @@
         </dependency>
         <dependency>
             <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>plugin-features</artifactId>
+            <artifactId>grpc-features</artifactId>
             <version>${project.version}</version>
             <classifier>features</classifier>
             <type>xml</type>
         </dependency>
         <dependency>
             <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>taskset-features</artifactId>
+            <artifactId>snmp-plugin</artifactId>
             <version>${project.version}</version>
             <classifier>features</classifier>
             <type>xml</type>
@@ -51,20 +48,6 @@
             <classifier>features</classifier>
             <type>xml</type>
         </dependency>
-        <dependency>
-            <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>icmp-plugin</artifactId>
-            <version>${project.version}</version>
-            <classifier>features</classifier>
-            <type>xml</type>
-        </dependency>
-        <dependency>
-            <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>traps-features</artifactId>
-            <version>${project.version}</version>
-            <classifier>features</classifier>
-            <type>xml</type>
-        </dependency>
     </dependencies>
 
     <build>
@@ -72,17 +55,19 @@
             <plugin>
                 <groupId>org.apache.karaf.tooling</groupId>
                 <artifactId>karaf-maven-plugin</artifactId>
+                <extensions>true</extensions>
                 <executions>
                     <execution>
-                        <id>default-verify</id>
+                        <id>verify</id>
+                        <goals>
+                            <goal>verify</goal>
+                        </goals>
+                        <phase>verify</phase>
                         <configuration>
                             <descriptors>
                                 <descriptor>mvn:org.apache.karaf.features/framework/${karaf.version}/xml/features</descriptor>
                                 <descriptor>mvn:org.apache.karaf.features/standard/${karaf.version}/xml/features</descriptor>
-                                <descriptor>mvn:org.apache.karaf.features/spring/${karaf.version}/xml/features</descriptor>
-                                <descriptor>mvn:org.apache.cxf.karaf/apache-cxf/${cxf.version}/xml/features</descriptor>
-                                <descriptor>mvn:org.apache.camel.karaf/apache-camel/${camel.version}/xml/features</descriptor>
-                                <descriptor>mvn:org.apache.aries.jax.rs/org.apache.aries.jax.rs.features/${aries.jax.rs.whiteboard.version}/xml</descriptor>
+                                <descriptor>mvn:org.opennms.horizon.minion/snmp-plugin/${project.version}/xml/features</descriptor>
                                 <descriptor>file:${project.build.directory}/feature/feature.xml</descriptor>
                             </descriptors>
                             <javase>11</javase>
@@ -90,7 +75,7 @@
                                 <feature>framework</feature>
                             </framework>
                             <features>
-                                <feature>minion-ng*</feature>
+                                <feature>minion-traps</feature>
                             </features>
                         </configuration>
                     </execution>
@@ -98,4 +83,5 @@
             </plugin>
         </plugins>
     </build>
+
 </project>
