+++ minion/traps/features/src/main/feature/feature.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<features name="minion-ipc-${project.version}" xmlns="http://karaf.apache.org/xmlns/features/v1.6.0">
+
+    <repository>mvn:org.opennms.horizon.shared/snmp-api/${project.version}/xml/features</repository>
+    <repository>mvn:org.opennms.horizon.minion/snmp-plugin/${project.version}/xml/features</repository>
+    <repository>mvn:org.opennms.horizon.minion/ipc-features/${project.version}/xml/features</repository>
+    <repository>mvn:org.opennms.horizon.minion/grpc-features/${project.version}/xml/features</repository>
+    <repository>mvn:org.opennms.horizon.minion/observability-features/${project.version}/xml/features</repository>
+
+    <feature name="minion-traps" version="${project.version}">
+        <feature>ipc-minion</feature>
+        <feature>snmp-snmp4j</feature>
+        <bundle>mvn:org.opennms.horizon.minion/traps-listener/${project.version}</bundle>
+    </feature>
+
+</features>
