+++ minion/traps/listener/pom.xml
@@ -3,14 +3,14 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>minion</artifactId>
+        <artifactId>traps</artifactId>
         <groupId>org.opennms.horizon.minion</groupId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>traps-listener</artifactId>
-    <name>OpenNMS Horizon :: Minion :: Trap :: Listener</name>
+    <name>OpenNMS Horizon :: Minion :: Traps :: Listener</name>
     <packaging>bundle</packaging>
 
     <dependencies>
