+++ minion/traps/listener/src/main/java/org/opennms/horizon/minion/traps/listener/TrapListener.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.minion.traps;
+package org.opennms.horizon.minion.traps.listener;
 
 import com.google.protobuf.ByteString;
 import org.opennms.horizon.grpc.traps.contract.TrapDTO;
